tts/requirements.txt
coqui-tts==0.27.1
fastapi
uvicorn[standard]
numpy
